refactor: split tests into separate files
